UserAndBlogServiceUpdate

+++ service/src/main/java/com/mpls/v2/service/impl/BlogServiceImpl.java
@@ -3,18 +3,25 @@ package com.mpls.v2.service.impl;
 import com.mpls.v2.model.Blog;
 import com.mpls.v2.repository.BlogRepository;
 import com.mpls.v2.service.BlogService;
+import com.mpls.v2.service.exceptions.SaveException;
 import org.springframework.beans.factory.annotation.Autowired;
+
 
 import java.util.List;
 
-public class BlogServiceImpl implements BlogService{
+public class BlogServiceImpl implements BlogService {
 
     @Autowired
     BlogRepository blogRepository;
 
     @Override
     public Blog save(Blog blog) {
-        return blogRepository.save(blog);
+        if (blog != null) {
+            return blogRepository.save(blog);
+        } else {
+            throw new SaveException("blog muast be not null");
+        }
+
     }
 
     @Override
@@ -24,12 +31,20 @@ public class BlogServiceImpl implements BlogService{
 
     @Override
     public Blog findOne(Long id) {
-        return blogRepository.findOne(id);
+        if (id != null || id >= 0) {
+            return blogRepository.findOne(id);
+        } else {
+            throw new SaveException("id must be not null");
+        }
     }
 
     @Override
     public Boolean delete(Long id) {
-        blogRepository.delete(blogRepository.findOne(id));
-        return true;
+        if (id != null || id >= 0) {
+            blogRepository.delete(blogRepository.findOne(id));
+            return true;
+        } else {
+            throw new SaveException("id must be not null");
+        }
     }
 }
